# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Market API
  version: 1.0.0
  description: |
    Backend API cho **Smart Market – Chợ thông minh** (Zalo Mini App + CMS).

    ## Xác thực
    - **Mini app**: `POST /auth/zalo-login` với access token Zalo. Môi trường dev bật `ZALO_MOCK=true`
      cho phép token dạng `mock:<zaloId>:<Tên>` (vd `mock:zalo-buyer-1:Người Mua`).
    - **CMS**: `POST /auth/login` với username/email/số điện thoại + mật khẩu.
    - Sau khi đăng nhập, bấm **Authorize** và dán `accessToken` (không cần chữ "Bearer").

    ## Vai trò & phạm vi
    | Role | Phạm vi |
    |---|---|
    | super_admin | Toàn hệ thống |
    | province_admin | Các chợ trong tỉnh được gán (chủ yếu chỉ đọc + escalate phản ánh) |
    | market_manager | CRUD trong các chợ được gán |
    | user | Người mua; nếu `merchant_status=active` và có sạp → dùng được nhóm `/seller/*` |

    ## Quy ước
    - Envelope: `{ success, data, meta? }` hoặc `{ success: false, error: { code, message, details? } }`.
    - Phân trang: `?page=1&limit=20` (limit tối đa 100) → `meta { page, limit, total, totalPages }`.
    - Tiền VND số nguyên. Thời gian ISO 8601. Key JSON camelCase.
    - Tài khoản demo (seed): SA `admin@smartmarket.vn / Admin@12345`; MM `mm.trungtam@smartmarket.vn / Manager@123`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Search:
      name: search
      in: query
      schema: { type: string }
    MarketIdQuery:
      name: marketId
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: STATE_INVALID }
            message: { type: string, example: Đơn đang ở trạng thái "completed", không thể chuyển sang "confirmed" }
            details: { nullable: true }
...
  responses:
    Ok:
      description: Thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: {}
    Paginated:
      description: Danh sách có phân trang
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: { type: array, items: {} }
              meta:
                type: object
                properties:
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  total: { type: integer, example: 135 }
                  totalPages: { type: integer, example: 7 }
    Err:
      description: Lỗi (400/401/403/404/409/429 — xem `error.code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security: []
paths:
...
  /markets/{id}/traders:
    get:
      tags: [Markets (App)]
      summary: Danh sách tiểu thương/sạp đang kinh doanh trong chợ
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - { name: zoneId, in: query, schema: { type: string, format: uuid } }
        - { name: hasPromotion, in: query, schema: { type: boolean } }
        - { name: minRating, in: query, schema: { type: number } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
...
  /admin/traders:
    get:
      tags: [Admin - Traders]
      summary: Danh sách tiểu thương theo scope
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: status, in: query, schema: { type: string, enum: [active, suspended, inactive] } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /admin/traders/{id}:
    get:
      tags: [Admin - Traders]
      summary: Chi tiết tiểu thương (hợp đồng, sạp, SP, đơn, phản ánh gần nhất)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /admin/traders/{id}/status:
    patch:
      tags: [Admin - Traders]
      summary: Đổi trạng thái (suspended sẽ ẩn sản phẩm + chặn quyền seller)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, suspended, inactive] }
                reason: { type: string, nullable: true }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }

  # ================= ADMIN - PRODUCTS =================
...
  /admin/reports/traders:
    get:
      tags: [Admin - Reports]
      summary: Top tiểu thương doanh thu cao + bị phản ánh nhiều
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/MarketIdQuery'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...